Added Packages
I added a GUI package for all GUI classes and a Model packages for our other classes
