+++ AsteroidMining/Asteroid Mining/src/com/InProgress/GUI/LeaveResourcesWindow.java
@@ -1,4 +1,4 @@
-package com.InProgress;
+package com.InProgress.GUI;
 
 public class LeaveResourcesWindow extends javax.swing.JFrame {
