+++ AsteroidMining/Asteroid Mining/src/com/InProgress/Model/Asteroid.java
@@ -1,6 +1,5 @@
-package com.InProgress;
+package com.InProgress.Model;
 
-import java.io.FileNotFoundException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Random;
