+++ AsteroidMining/Asteroid Mining/src/com/InProgress/Model/Carbon.java
@@ -1,4 +1,4 @@
-package com.InProgress;
+package com.InProgress.Model;
 
 public class Carbon extends ResourceBase {
