+++ AsteroidMining/Asteroid Mining/src/com/InProgress/Model/GameWindow.java
@@ -1,4 +1,4 @@
-package com.InProgress;
+package com.InProgress.Model;
 
 public class GameWindow extends javax.swing.JFrame {
