+++ AsteroidMining/Asteroid Mining/src/com/InProgress/Model/Inventory.java
@@ -1,4 +1,4 @@
-package com.InProgress;
+package com.InProgress.Model;
 
 import java.util.ArrayList;
 import java.util.List;
