+++ AsteroidMining/Asteroid Mining/src/com/InProgress/Model/Iron.java
@@ -1,4 +1,4 @@
-package com.InProgress;
+package com.InProgress.Model;
 
 public class Iron extends ResourceBase {
