+++ AsteroidMining/Asteroid Mining/src/com/InProgress/Model/Main.java
@@ -1,9 +1,6 @@
-package com.InProgress;
+package com.InProgress.Model;
 
-import javax.swing.*;
-import java.io.BufferedReader;
 import java.io.IOException;
-import java.io.InputStreamReader;
 
 
 public class Main {
