+++ AsteroidMining/Asteroid Mining/src/com/InProgress/Model/Robot.java
@@ -1,4 +1,4 @@
-package com.InProgress;
+package com.InProgress.Model;
 
 import java.util.Random;
