+++ AsteroidMining/Asteroid Mining/src/com/InProgress/Model/Settler.java
@@ -1,4 +1,4 @@
-package com.InProgress;
+package com.InProgress.Model;
 
 public class Settler extends TravellerBase {
