+++ AsteroidMining/Asteroid Mining/src/com/InProgress/Model/Sun.java
@@ -1,4 +1,4 @@
-package com.InProgress;
+package com.InProgress.Model;
 
 
 import java.util.Random;
