+++ AsteroidMining/Asteroid Mining/src/com/InProgress/Model/TransportGate.java
@@ -1,4 +1,4 @@
-package com.InProgress;
+package com.InProgress.Model;
 
 public class TransportGate {
