+++ AsteroidMining/Asteroid Mining/src/com/InProgress/Model/Uranium.java
@@ -1,4 +1,4 @@
-package com.InProgress;
+package com.InProgress.Model;
 
 public class Uranium extends ResourceBase{
