+++ AsteroidMining/Asteroid Mining/src/com/InProgress/Model/WaterIce.java
@@ -1,5 +1,4 @@
-package com.InProgress;
-import com.InProgress.Asteroid;
+package com.InProgress.Model;
 
 public class WaterIce extends ResourceBase{
